user balance has been changed to int

+++ src/main/java/KBD/Config.java
@@ -4,6 +4,6 @@ public class Config {
     public static final String NO_IMAGE_PATH = "/KhaneBeDoosh/assets/no-pic.jpg";
     public static final String API_KEY = "60ffead0-12e7-11e8-87b4-496f79ef1988";
     public static final String BANK_URL = "http://acm.ut.ac.ir/ieBank/pay";
-    public static final float HOUSE_OWNER_NUMBER_PRICE = 1000;
+    public static final int HOUSE_OWNER_NUMBER_PRICE = 1000;
     public static final int SUCCESS_RESPONSE_STATUS_CODE = 200;
 }
